refactor: move the translation check into the country configs
