+++ packages/testland/package.json
@@ -27,7 +27,7 @@
     "start": "pnpm setup-analytics && pnpm setup-reference-data &&  cross-env NODE_ENV=development NODE_OPTIONS=--dns-result-order=ipv4first nodemon --watch src --exec tsx src/index.ts",
     "start:prod": "tsx src/index.ts",
     "metabase": "bash assets/metabase/run-dev.sh",
-    "extract:translations": "tsx ../toolkit/src/cli.ts check-translations",
+    "extract:translations": "tsx src/check-translations.ts",
     "sort-translations": "cross-env NODE_ENV=development tsx src/sort-translations.ts"
   },
   "devDependencies": {
@@ -61,6 +61,7 @@
     "playwright-ctrf-json-reporter": "^0.0.18",
     "prettier": "3.5.3",
     "react-intl": "^6.4.3",
+    "ts-morph": "^27.0.2",
     "vite": "^8.1.5",
     "vite-tsconfig-paths": "^6.1.1",
     "vitest": "^4.1.10"
